add login service check

+++ gameregistry/src/main/groovy/es/us/dad/gameregistry/controller/SessionController.groovy
@@ -1,6 +1,7 @@
 package es.us.dad.gameregistry.controller
 
 import es.us.dad.gameregistry.domain.GameSession
+import es.us.dad.gameregistry.service.LoginService
 import es.us.dad.gameregistry.service.SessionService
 import es.us.dad.gameregistry.util.DELETE
 import es.us.dad.gameregistry.util.GET
@@ -14,12 +15,17 @@ class SessionController extends RestController {
     // TODO: dependency injection
     private final SessionService sessionService
 
-    public SessionController(SessionService sessionService) {
+    public SessionController(LoginService loginService, SessionService sessionService) {
+        super(loginService)
         this.sessionService = sessionService
     }
 
     @GET("/sessions/:id")
     public void getSession(HttpServerRequest request) {
+        if (!validateUserAuthentication(request)) {
+            return
+        }
+
         UUID id = UUID.fromString(request.params.get("id"))
         GameSession session = sessionService.getSession(id)
 
@@ -36,6 +42,10 @@ class SessionController extends RestController {
 
     @PUT("/sessions/:id")
     public void changeSession(HttpServerRequest request) {
+        if (!validateUserAuthentication(request)) {
+            return
+        }
+
         UUID id = UUID.fromString(request.params.get("id"))
         GameSession session = sessionService.finishSession(id)
 
@@ -47,6 +57,10 @@ class SessionController extends RestController {
 
     @DELETE("/sessions/:id")
     public void deleteSession(HttpServerRequest request) {
+        if (!validateUserAuthentication(request)) {
+            return
+        }
+
         UUID id = UUID.fromString(request.params.get("id"))
         boolean success = sessionService.deleteSession(id)
